Major changes to code
Will show color based on price movement compared to the previous return. If the price has increased, will turn green, and vice versa. If the price remains as is, will keep the same color as previous return. Also added a net difference counter, which compares the returned price to the first returned price when the program is launched in order to keep track of movement within the session.